fix: cleanup unused files
